Update Bot
Moved Commands out of bot.py to keep it neat, and moved the commands to a different folder in the Cog called Commands.py
