--- cogs/example.py
@@ -1,6 +1,0 @@
-import Discord
-from discord.ext import commands
-
-class Example(commands.Cog):
-    def __init__(self, client):
-        self.client = client
